Materials and dpa reworked
