+++ irradiapy/materials/o.py
@@ -1,0 +1,5 @@
+"""O material."""
+
+from irradiapy.materials.material import Material
+
+O = Material(atomic_number=8, mass_number=15.999)
